Replace `textarea` with `div contenteditable`
Chrome has a new bug where any ZWNJ is replaced by a dotted circle when
the font uses USE for that item and the element is styled with a
space-preserving value of `white-space`. `textarea` uses `white-space:
break-spaces` by default and it seems not to be overridable. I therefore
had to reimplement it with a `div contenteditable`, which is not at all
straightforward.

This commit also fixes some bugs. (It apparently wasn’t straightforward
to get a `textarea` working either.)