Scroll the text box only when necessary
Calling `scrollIntoView` on every new character is slow. To avoid
unnecessary reflow, it now only scrolls when it detects that it might
need to. It sometimes scrolls unnecessarily, in that it always scrolls
such that the new character is at the bottom of the text box even if it
was visible anyway, but it never scrolls incorrectly. There may be a
delay between typing a character and scrolling.

This makes typing feel slightly more responsive, but it is still
sluggish. The real bottleneck is the font itself.